[db] Switch to string primary key for parliaments
The abbreviation of each parliament is supposed to be unique and can then be directly used as a primary key instead of relying on an auto-increment field.

Moreover, using the string code as an ID gives the additional advantage of making some data relationships more immediately understandable.

This commit also brings the required changes to the related seeder.